Move sortExceptions to tests/helpers folder
