+++ packages/core/comparison/lib/comparePhase1.ts
@@ -17,8 +17,8 @@ import generateMockPncQueryResultFromAho from "../../tests/helpers/comparison/ge
 import getPncQueryTimeFromAho from "../../tests/helpers/comparison/getPncQueryTimeFromAho"
 import parseIncomingMessage from "../../tests/helpers/comparison/parseIncomingMessage"
 import MockPncGateway from "../../tests/helpers/MockPncGateway"
+import sortExceptions from "../../tests/helpers/sortExceptions"
 import isIntentionalDifference from "./isIntentionalDifference"
-import { sortExceptions } from "./sortExceptions"
 import { sortTriggers } from "./sortTriggers"
 import { matchingExceptions } from "./summariseMatching"
 import { xmlOutputDiff, xmlOutputMatches } from "./xmlOutputComparison"
