+++ packages/core/tsconfig.build.json
@@ -9,6 +9,7 @@
     "phase3/tests/helpers/getPncErrorMessages.ts",
     "phase3/tests/helpers/getPncOperationsFromPncUpdateDataset.ts",
     "tests/helpers/MockPncGateway.ts",
+    "tests/helpers/sortExceptions.ts",
     "tests/helpers/comparison/generateMockPncQueryResultFromAho.ts",
     "tests/helpers/comparison/getPncQueryTimeFromAho.ts",
     "tests/helpers/comparison/parseIncomingMessage.ts",
